Adding the method update to the userController in order to update users and some more modifications

+++ src/main/java/com/manuelsarante/ApiBeBank/web/BankingAccountController.java
@@ -7,6 +7,7 @@ import com.manuelsarante.ApiBeBank.domain.User;
 import com.manuelsarante.ApiBeBank.dto.UpdateAmountAccountDto;
 import com.manuelsarante.ApiBeBank.service.BankingAccountService;
 import com.manuelsarante.ApiBeBank.service.CardsService;
+import com.manuelsarante.ApiBeBank.service.UserService;
 import com.manuelsarante.ApiBeBank.specialfunctions.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +15,10 @@ import org.springframework.web.bind.annotation.*;
 import java.time.LocalDate;
 import java.util.List;
 
-
+/*
+ * Copyright (c) Manuel Antonio Sarante Sanchez 2023
+ * All rights reserved.
+ */
 @RestController
 @RequestMapping("/account")
 public class BankingAccountController {
@@ -25,10 +29,14 @@ public class BankingAccountController {
     @Autowired
     CardsService cardsService;
 
+    @Autowired
+    UserService userService;
+
 
     //With this method you only have to set the id of the user you want to add a new account in the end point and in the body you set the amount you are going to have in the account
     @PostMapping("/save/{id}")
     public ResponseEntity<?> saveBankingAccount(@PathVariable Long id, @RequestBody UpdateAmountAccountDto updateAmountAccountDto){
+
         //First generate the banking account, the card number, the expire date of the card and Cvv of the card
         String accountNumber = new AccountNumber().generateNumber();
         String cardNumber = new LuhnNumber().generate();
@@ -51,27 +59,34 @@ public class BankingAccountController {
                 break;
             }
         }
-        //Create the user
-        User user = new User();
-        //to the user we only set the id of the user which the account is gonna be related with
-        user.setIdUser(id);
-        //Create the bankingAccount
-        BankingAccount bankingAccountToSave = new BankingAccount();
-        //Create the card
-        Cards cardToSave = new Cards();
-        //Set the values to the card
-        cardToSave.setCardNumber(cardNumber);
-        cardToSave.setCreationDate(LocalDate.now());
-        cardToSave.setExpireDate(expireDate);
-        cardToSave.setCvv(Integer.parseInt(cvv));
-        //Set The values to the banking account
-        bankingAccountToSave.setAccountNumber(accountNumber);
-        bankingAccountToSave.setUser(user);
-        bankingAccountToSave.setMountAccount(updateAmountAccountDto.getMount());
-        bankingAccountToSave.setCards(cardToSave);
+        //find the user in database
+        User userFound = userService.findById(id);
+        //If user is different from null it will save the account
+        if(userFound!=null){
+            //Create the user
+            User user = new User();
+            //to the user we only set the id of the user which the account is gonna be related with
+            user.setIdUser(id);
+            //Create the bankingAccount
+            BankingAccount bankingAccountToSave = new BankingAccount();
+            //Create the card
+            Cards cardToSave = new Cards();
+            //Set the values to the card
+            cardToSave.setCardNumber(cardNumber);
+            cardToSave.setCreationDate(LocalDate.now());
+            cardToSave.setExpireDate(expireDate);
+            cardToSave.setCvv(Integer.parseInt(cvv));
+            //Set The values to the banking account
+            bankingAccountToSave.setAccountNumber(accountNumber);
+            bankingAccountToSave.setUser(user);
+            bankingAccountToSave.setMountAccount(updateAmountAccountDto.getMount());
+            bankingAccountToSave.setCards(cardToSave);
 
-        bankingAccountService.insert(bankingAccountToSave);
-        return ResponseEntity.ok(bankingAccountToSave);
+            bankingAccountService.insert(bankingAccountToSave);
+            return ResponseEntity.ok(bankingAccountToSave);
+        }else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PutMapping("/updateMount/{id}")
@@ -102,7 +117,12 @@ public class BankingAccountController {
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deletebankingAccount(@PathVariable Long id){
         BankingAccount bankingAccount = bankingAccountService.findByIdAccount(id);
-        bankingAccountService.delete(bankingAccount);
-        return ResponseEntity.ok(new Messages("Banking account deleted"));
+        if(bankingAccount!=null){
+            bankingAccountService.delete(bankingAccount);
+            return ResponseEntity.ok(new Messages("Banking account deleted"));
+        }else{
+            return ResponseEntity.notFound().build();
+        }
+
     }
 }
